+++ src/serialization/string_utils.cpp
@@ -49,9 +49,7 @@ bool isnewline(const char c)
 bool portable_isspace(const char c)
 {
 	// returns true only on ASCII spaces
-	if (static_cast<unsigned char>(c) >= 128)
-		return false;
-	return isnewline(c) || isspace(static_cast<unsigned char>(c));
+	return c == '\r' || c == '\n' || c == ' ' || c == '\t' || c == '\v' || c == '\f';
 }
 
 // Make sure we regard '\r' and '\n' as a space, since Mac, Unix, and DOS